# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type CheckoutInputItemsItem = {
  productId: string;
  /**
   * @minimum 1
   * @maximum 50
   */
  quantity: number;
};

export interface CheckoutInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  /** DNI o CUIT del comprador (opcional, sólo dígitos, 7-11 caracteres) */
  customerDni?: string | null;
  /**
   * @minItems 1
   * @maxItems 50
   */
  items: CheckoutInputItemsItem[];
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  imagePath: string;
  /** @nullable */
  filePath: string | null;
  isBestSeller: boolean;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  id: string;
  name: string;
  category: string;
  /** @minimum 0 */
  price: number;
  imagePath: string;
  /** @nullable */
  filePath?: string | null;
  isBestSeller?: boolean;
  isPublished?: boolean;
}

export interface ProductUpdate {
  name?: string;
  category?: string;
  /** @minimum 0 */
  price?: number;
  imagePath?: string;
  /** @nullable */
  filePath?: string | null;
  isBestSeller?: boolean;
  isPublished?: boolean;
}

export interface OrderItem {
  productId: string;
  name: string;
  price: number;
  quantity: number;
  imagePath: string;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  mercadopago: "mercadopago",
  uala: "uala",
  paypal: "paypal",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
  refunded: "refunded",
} as const;

export type OrderConfirmationSource =
  | (typeof OrderConfirmationSource)[keyof typeof OrderConfirmationSource]
  | null;

export const OrderConfirmationSource = {
  webhook: "webhook",
  manual: "manual",
  "paypal-capture": "paypal-capture",
} as const;

export interface Order {
  id: string;
  invoiceNumber: number;
  customerName: string;
  customerEmail: string;
  customerDni?: string | null;
  items: OrderItem[];
  total: number;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  confirmationSource?: OrderConfirmationSource;
  createdAt: string;
}

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  mercadopago: "mercadopago",
  uala: "uala",
  paypal: "paypal",
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerDni?: string | null;
  /** @minItems 1 */
  items: OrderItem[];
  paymentMethod: OrderInputPaymentMethod;
}

export type OrderStatsRevenueByDayItem = {
  date: string;
  revenue: number;
  orders: number;
};

export type OrderStatsTopProductsItem = {
  productId: string;
  name: string;
  quantity: number;
  revenue: number;
};

export type OrderStatsRevenueByMethodItem = {
  method: string;
  revenue: number;
  orders: number;
};

export interface OrderStats {
  totalRevenue: number;
  totalOrders: number;
  revenueToday: number;
  revenueThisWeek: number;
  revenueThisMonth: number;
  ordersToday: number;
  revenueByDay: OrderStatsRevenueByDayItem[];
  topProducts: OrderStatsTopProductsItem[];
  revenueByMethod: OrderStatsRevenueByMethodItem[];
}

export type RequestUploadUrlBody = {
  name: string;
  size: number;
  contentType: string;
};

export type RequestUploadUrl200 = {
  uploadURL: string;
  objectPath: string;
};

export type CreateMercadoPagoPreference200 = {
  init_point: string;
  sandbox_init_point?: string;
  orderId: string;
};

export type CreatePaypalOrder200 = {
  ppOrderId: string;
  orderId: string;
};

export type CapturePaypalOrderBody = {
  ppOrderId: string;
  orderId: string;
};

export type GetUalaLink200 = {
  url: string;
};

export type GetOrderInvoice200 = {
  invoiceNumber: number;
};