# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create manual order (admin-only — requireAdmin)
      description: >
        Reserved for admin manual order entry. Real customer orders must go through
        /payments/mercadopago/preference or /payments/paypal/create-order and are
        confirmed by the gateway before being marked paid.
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not in admin allowlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/stats:
    get:
      operationId: getOrderStats
      tags: [orders]
      summary: Aggregate sales metrics (admin)
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}/invoice:
    get:
      operationId: getOrderInvoice
      tags: [orders]
      summary: Get invoice number for a confirmed order (public)
      description: >
        Returns only the invoiceNumber for a paid order. Safe to expose publicly —
        orderId is a UUID from MP external_reference and no sensitive data is returned.
        Used by checkout/success to show invoice number after MP redirect.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice number
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceNumber:
                    type: integer
                required: [invoiceNumber]
        "404":
          description: Order not found or not paid
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: integer
        quantity:
          type: integer
        imagePath:
          type: string
      required:
        - productId
        - name
        - price
        - quantity
        - imagePath
    Order:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerDni:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: integer
        paymentMethod:
          type: string
          enum: [mercadopago, uala, paypal]
        status:
          type: string
          enum: [pending, paid, failed, refunded]
        confirmationSource:
          type: string
          enum: [webhook, manual, paypal-capture]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - customerName
        - customerEmail
        - items
        - total
        - paymentMethod
        - status
        - createdAt
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          format: email
        customerDni:
          type: string
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        paymentMethod:
          type: string
          enum: [mercadopago, uala, paypal]
      required:
        - customerName
        - customerEmail
        - items
        - paymentMethod
    OrderStats:
      type: object
      properties:
        totalRevenue:
          type: integer
        totalOrders:
          type: integer
        revenueToday:
          type: integer
        revenueThisWeek:
          type: integer
        revenueThisMonth:
          type: integer
        ordersToday:
          type: integer
        revenueByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: integer
              orders:
                type: integer
            required: [date, revenue, orders]
        topProducts:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              name:
                type: string
              quantity:
                type: integer
              revenue:
                type: integer
            required: [productId, name, quantity, revenue]
        revenueByMethod:
          type: array
          items:
            type: object
            properties:
              method:
                type: string
              revenue:
                type: integer
              orders:
                type: integer
            required: [method, revenue, orders]
      required:
        - totalRevenue
        - totalOrders
        - revenueToday
        - revenueThisWeek
        - revenueThisMonth
        - ordersToday
        - revenueByDay
        - topProducts
        - revenueByMethod