# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderConfirmationSource } from "./orderConfirmationSource";
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  invoiceNumber: number;
  customerName: string;
  customerEmail: string;
  customerDni?: string | null;
  items: OrderItem[];
  total: number;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  confirmationSource?: OrderConfirmationSource;
  createdAt: Date;
}