# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from "./orderInputPaymentMethod";
import type { OrderItem } from "./orderItem";

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerDni?: string | null;
  /** @minItems 1 */
  items: OrderItem[];
  paymentMethod: OrderInputPaymentMethod;
}